+++ tests/sparseml/transformers/sparsification/test_compress_tensor_utils.py
@@ -20,8 +20,18 @@ import torch
 from transformers import AutoConfig
 
 import sparseml.core.session as session_manager
-from compressed_tensors import SPARSITY_CONFIG_NAME
+from compressed_tensors import (
+    COMPRESSION_CONFIG_NAME,
+    QUANTIZATION_CONFIG_NAME,
+    SPARSITY_CONFIG_NAME,
+)
 from compressed_tensors.config import BitmaskConfig, DenseSparsityConfig
+from compressed_tensors.quantization import (
+    QuantizationStatus,
+    compress_quantized_weights,
+    freeze_module_quantization,
+)
+from safetensors import safe_open
 from sparseml.transformers import SparseAutoModelForCausalLM, oneshot
 from sparseml.transformers.compression.sparsity_config import SparsityConfigMetadata
 
@@ -59,6 +69,7 @@ def test_sparse_model_reload(compressed, config, dtype, tmp_path):
         splits=splits,
         oneshot_device=device,
         precision=dtype,
+        clear_sparse_session=False,
     )
 
     model = SparseAutoModelForCausalLM.from_pretrained(
@@ -77,7 +88,8 @@ def test_sparse_model_reload(compressed, config, dtype, tmp_path):
     )
 
     config = AutoConfig.from_pretrained(tmp_path / "compress_out")
-    sparsity_config = getattr(config, SPARSITY_CONFIG_NAME, None)
+    compression_config = getattr(config, COMPRESSION_CONFIG_NAME, None)
+    sparsity_config = compression_config.get(SPARSITY_CONFIG_NAME, None)
     assert (
         sparsity_config["format"] == "dense"
         if (not compressed and config is None)
@@ -127,5 +139,139 @@ def test_dense_model_save(tmp_path, skip_compression_stats, save_compressed):
     config = AutoConfig.from_pretrained(tmp_path / "dense_out")
     sparsity_config = getattr(config, SPARSITY_CONFIG_NAME, None)
     assert sparsity_config is None
+
+    shutil.rmtree(tmp_path)
+
+
+@pytest.mark.parametrize(
+    "format,dtype",
+    [
+        ["dense", torch.float32],
+        ["dense", torch.float16],
+        ["int_quantized", torch.float32],
+        # [True, "int_quantized", torch.float16],
+    ],
+)
+def test_quant_model_reload(format, dtype, tmp_path):
+    recipe_str = "tests/sparseml/transformers/compression/recipes/new_quant_simple.yaml"
+    model_path = "Xenova/llama2.c-stories15M"
+    device = "cuda:0"
+    if not torch.cuda.is_available():
+        device = "cpu"
+    dataset = "open_platypus"
+    concatenate_data = False
+    num_calibration_samples = 64
+    output_dir = tmp_path / "oneshot_out"
+    splits = {"calibration": "train[:10%]"}
+
+    # create a quantized model
+    oneshot(
+        model=model_path,
+        dataset=dataset,
+        output_dir=output_dir,
+        num_calibration_samples=num_calibration_samples,
+        recipe=recipe_str,
+        concatenate_data=concatenate_data,
+        splits=splits,
+        oneshot_device=device,
+        precision=dtype,
+    )
+
+    model = SparseAutoModelForCausalLM.from_pretrained(
+        tmp_path / "oneshot_out", torch_dtype=dtype
+    )
+
+    for _, module in model.named_modules():
+        if hasattr(module, "quantization_scheme"):
+            assert module.weight.dtype == dtype
+            assert module.quantization_status == QuantizationStatus.FROZEN
+
+    model.save_pretrained(
+        tmp_path / "compress_out",
+        quantization_format=format,
+        save_compressed=True,
+    )
+
+    config = AutoConfig.from_pretrained(tmp_path / "compress_out")
+    compression_config = getattr(config, COMPRESSION_CONFIG_NAME, None)
+    quant_config = compression_config.get(QUANTIZATION_CONFIG_NAME, None)
+    assert quant_config["format"] == format
+
+    dense_model = SparseAutoModelForCausalLM.from_pretrained(
+        tmp_path / "compress_out", torch_dtype="auto"
+    )
+
+    og_state_dict = model.state_dict()
+    reconstructed_state_dict = dense_model.state_dict()
+    assert len(og_state_dict) == len(reconstructed_state_dict)
+    for key in og_state_dict.keys():
+        dense_tensor = og_state_dict[key]
+        reconstructed_tensor = reconstructed_state_dict[key]
+        assert dense_tensor.dtype == reconstructed_tensor.dtype
+        if key.endswith("weight") and format != "dense":
+            # we don't expect an exact match for compressed
+            diff = torch.abs(dense_tensor - reconstructed_tensor)
+            assert not torch.any(diff > 0.01).item()
+        else:
+            assert torch.equal(dense_tensor, reconstructed_tensor)
+
+    shutil.rmtree(tmp_path)
+
+
+@pytest.mark.parametrize(
+    "status,expected_format,expected_dtype",
+    [
+        [QuantizationStatus.FROZEN, "dense", torch.float32],
+        [QuantizationStatus.COMPRESSED, "int-quantized", torch.int8],
+    ],
+)
+def test_quant_infer_format(status, expected_format, expected_dtype, tmp_path):
+    recipe_str = "tests/sparseml/transformers/compression/recipes/new_quant_simple.yaml"
+    model_path = "Xenova/llama2.c-stories15M"
+    device = "cuda:0"
+    if not torch.cuda.is_available():
+        device = "cpu"
+    dataset = "open_platypus"
+    concatenate_data = False
+    num_calibration_samples = 64
+    output_dir = tmp_path / "oneshot_out"
+    splits = {"calibration": "train[:10%]"}
+
+    model = SparseAutoModelForCausalLM.from_pretrained(model_path)
+
+    # create a quantized model
+    oneshot(
+        model=model,
+        dataset=dataset,
+        output_dir=output_dir,
+        num_calibration_samples=num_calibration_samples,
+        recipe=recipe_str,
+        concatenate_data=concatenate_data,
+        splits=splits,
+        oneshot_device=device,
+    )
+
+    if status == QuantizationStatus.FROZEN:
+        model.apply(freeze_module_quantization)
+    elif status == QuantizationStatus.COMPRESSED:
+        model.apply(compress_quantized_weights)
+
+    for _, module in model.named_modules():
+        if hasattr(module, "quantization_scheme"):
+            assert module.quantization_status == status
+
+    model.save_pretrained(tmp_path / "compress_out")
+
+    config = AutoConfig.from_pretrained(tmp_path / "compress_out")
+    compression_config = getattr(config, COMPRESSION_CONFIG_NAME, None)
+    quant_config = compression_config.get(QUANTIZATION_CONFIG_NAME, None)
+    assert quant_config["quantization_status"] == status.value
+    assert quant_config["format"] == expected_format
+
+    with safe_open(
+        tmp_path / "compress_out" / "model.safetensors", framework="pt", device=device
+    ) as f:
+        test_tensor = f.get_tensor("model.layers.0.mlp.down_proj.weight")
+        assert test_tensor.dtype == expected_dtype
 
     shutil.rmtree(tmp_path)
